requirements.txt
py_ecc>=7.0.0
pycryptodome>=3.19
numpy>=1.24
matplotlib>=3.7
pandas>=2.0
scipy>=1.10
